feat: move color to utils.py
