+++ sebs/utils.py
@@ -5,6 +5,9 @@ import shutil
 import subprocess
 import sys
 import uuid
+import click
+import datetime
+
 from typing import List, Optional, TextIO, Union
 
 PROJECT_DIR = os.path.join(os.path.dirname(os.path.realpath(__file__)), os.pardir)
@@ -200,3 +203,20 @@ def catch_interrupt():
         sys.exit(signal.SIGINT)
 
     signal.signal(signal.SIGINT, handler)
+
+
+
+class Colors:
+    SUCCESS = "\033[92m"
+    STATUS = "\033[94m"
+    WARNING = "\033[93m"
+    ERROR = "\033[91m"
+    BOLD = "\033[1m"
+    END = "\033[0m"
+
+class ColoredPrinter:
+    @staticmethod
+    def print(color, logging_instance, message):
+        timestamp = datetime.datetime.now().strftime("%H:%M:%S")
+        logging_instance.info(message)
+        click.echo(f"{color}{Colors.BOLD}[{timestamp}]{Colors.END} {message}")
